refactor: model 내 패키지 수정
